Add in-game timer for speedruns
With the cvars cl_drawtotaltimer and cl_drawleveltimer, the total and
level in-game timers can be drawn on the hud. Higher values show more
non-integral digits. Additionally, at the end of a map, level and total
time are printed to the console and also stored in any demo stream, if
such exist. The output in the console will therefore be printed when
playing back the demo as well.