Revert "Fix biggun crash"
This reverts commit 9215a22bbcdb719b1ea8391b3e7dc48345fadab6.
That fix was hacky and did not solve the core of the problem. skullernet
pushed a proper fix with commit d3980a0f751f9161ee9bb490ca1d42c18dc40250
that has since been merged into the fork kugelrund/q2pro, which is why
the rudimentary fix is reverted now.